Update UsageVsWeather.ipynb to use full data set.
This actually changes some of the conclusions previously suggested in the
document.  Specifically, it no longer seems that temperature and computer usage
are at all correlated.

Replace scatter plot with hexbin plot due to massive number of points.

Change titles and axis labels to be more consistent and more explanatory.